+++ src/actions/index.js
@@ -1,16 +1,3 @@
-export const GET_TOKEN = "GET_TOKEN";
-const SET_TOKEN = "SET_TOKEN";
+import { GET_TOKEN, getToken, setToken } from './authActions'
 
-export const getToken = (code) => {
-    return {
-        type: "GET_TOKEN",
-        payload: code
-    }
-}
-
-export const setToken = data => {
-    return {
-        type: "SET_TOKEN",
-        payload: data
-    }
-}
+export { GET_TOKEN, getToken, setToken };
